refactor(iam): consolidate profile edit relations
