--- examples/banking/server/crates/iam/application/src/authorization/organization/organization_display_name_changer_relation.rs
@@ -1,18 +1,0 @@
-use appletheia::application::authorization::{Relation, RelationName, RelationRef, UsersetExpr};
-use appletheia::domain::Aggregate;
-
-use super::{Organization, OrganizationProfileEditorRelation};
-
-/// Allows organization profile editors to change an organization display name.
-pub struct OrganizationDisplayNameChangerRelation;
-
-impl Relation for OrganizationDisplayNameChangerRelation {
-    const REF: RelationRef = RelationRef::new(
-        Organization::TYPE,
-        RelationName::new("display_name_changer"),
-    );
-
-    const EXPR: UsersetExpr = UsersetExpr::ComputedUserset {
-        relation: OrganizationProfileEditorRelation::REF,
-    };
-}
