--- examples/banking/server/crates/iam/application/src/authorization/organization/organization_picture_changer_relation.rs
@@ -1,16 +1,0 @@
-use appletheia::application::authorization::{Relation, RelationName, RelationRef, UsersetExpr};
-use appletheia::domain::Aggregate;
-
-use super::{Organization, OrganizationProfileEditorRelation};
-
-/// Allows organization profile editors to change an organization picture.
-pub struct OrganizationPictureChangerRelation;
-
-impl Relation for OrganizationPictureChangerRelation {
-    const REF: RelationRef =
-        RelationRef::new(Organization::TYPE, RelationName::new("picture_changer"));
-
-    const EXPR: UsersetExpr = UsersetExpr::ComputedUserset {
-        relation: OrganizationProfileEditorRelation::REF,
-    };
-}
