--- examples/banking/server/crates/iam/application/src/authorization/organization/organization_website_url_changer_relation.rs
@@ -1,16 +1,0 @@
-use appletheia::application::authorization::{Relation, RelationName, RelationRef, UsersetExpr};
-use appletheia::domain::Aggregate;
-
-use super::{Organization, OrganizationProfileEditorRelation};
-
-/// Allows organization profile editors to change an organization website URL.
-pub struct OrganizationWebsiteUrlChangerRelation;
-
-impl Relation for OrganizationWebsiteUrlChangerRelation {
-    const REF: RelationRef =
-        RelationRef::new(Organization::TYPE, RelationName::new("website_url_changer"));
-
-    const EXPR: UsersetExpr = UsersetExpr::ComputedUserset {
-        relation: OrganizationProfileEditorRelation::REF,
-    };
-}
