--- examples/banking/server/crates/iam/application/src/authorization/user/user_picture_changer_relation.rs
@@ -1,15 +1,0 @@
-use appletheia::application::authorization::{Relation, RelationName, RelationRef, UsersetExpr};
-use appletheia::domain::Aggregate;
-
-use super::{User, UserProfileEditorRelation};
-
-/// Defines the `picture_changer` relation for `User`.
-pub struct UserPictureChangerRelation;
-
-impl Relation for UserPictureChangerRelation {
-    const REF: RelationRef = RelationRef::new(User::TYPE, RelationName::new("picture_changer"));
-
-    const EXPR: UsersetExpr = UsersetExpr::ComputedUserset {
-        relation: UserProfileEditorRelation::REF,
-    };
-}
