+++ examples/banking/server/crates/iam/application/src/command/user/user_bio_change/user_bio_change_command.rs
@@ -1,0 +1,11 @@
+use appletheia::command;
+use banking_iam_domain::{UserBio, UserId};
+use serde::{Deserialize, Serialize};
+
+/// Changes a user's bio.
+#[command(name = "user_bio_change")]
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+pub struct UserBioChangeCommand {
+    pub user_id: UserId,
+    pub bio: Option<UserBio>,
+}
