+++ examples/banking/server/crates/iam/application/src/command/user/user_bio_change/user_bio_change_command_handler_error.rs
@@ -1,0 +1,16 @@
+use appletheia::application::repository::RepositoryError;
+use banking_iam_domain::{User, UserError};
+use thiserror::Error;
+
+/// Represents errors returned while changing a user bio.
+#[derive(Debug, Error)]
+pub enum UserBioChangeCommandHandlerError {
+    #[error("user repository failed")]
+    UserRepository(#[from] RepositoryError<User>),
+
+    #[error("user aggregate failed")]
+    User(#[from] UserError),
+
+    #[error("user was not found")]
+    UserNotFound,
+}
