+++ examples/banking/server/crates/iam/application/src/command/user/user_bio_change/user_bio_change_output.rs
@@ -1,0 +1,5 @@
+use serde::{Deserialize, Serialize};
+
+/// Returned after a user bio change request is applied.
+#[derive(Clone, Debug, Default, PartialEq, Eq, Serialize, Deserialize)]
+pub struct UserBioChangeOutput;
