+++ examples/banking/server/crates/iam/application/src/command/user/user_display_name_change/user_display_name_change_output.rs
@@ -1,0 +1,5 @@
+use serde::{Deserialize, Serialize};
+
+/// Returned after a user display name change request is applied.
+#[derive(Clone, Debug, Default, PartialEq, Eq, Serialize, Deserialize)]
+pub struct UserDisplayNameChangeOutput;
