+++ examples/banking/server/crates/iam/application/src/command/user/user_picture_change.rs
@@ -1,103 +1,9 @@
-use appletheia::application::authorization::{
-    AuthorizationPlan, PrincipalRequirement, Relation, RelationshipRequirement,
-};
-use appletheia::application::command::{CommandHandled, CommandHandler};
-use appletheia::application::projection::{ProjectorDependencies, ProjectorSpec};
-use appletheia::application::repository::{Repository, RepositoryError};
-use appletheia::application::request_context::RequestContext;
-use appletheia::command;
-use banking_iam_domain::{User, UserError, UserId, UserPictureRef};
-use serde::{Deserialize, Serialize};
-use thiserror::Error;
+mod user_picture_change_command;
+mod user_picture_change_command_handler;
+mod user_picture_change_command_handler_error;
+mod user_picture_change_output;
 
-use crate::authorization::UserPictureChangerRelation;
-use crate::projection::UserOwnerRelationshipProjectorSpec;
-
-/// Changes a user's picture.
-#[command(name = "user_picture_change")]
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct UserPictureChangeCommand {
-    pub user_id: UserId,
-    pub picture: Option<UserPictureRef>,
-}
-
-/// Returned after a user picture change request is applied.
-#[derive(Clone, Debug, Default, PartialEq, Eq, Serialize, Deserialize)]
-pub struct UserPictureChangeOutput;
-
-/// Represents errors returned while changing a user picture.
-#[derive(Debug, Error)]
-pub enum UserPictureChangeCommandHandlerError {
-    #[error("user repository failed")]
-    UserRepository(#[from] RepositoryError<User>),
-
-    #[error("user aggregate failed")]
-    User(#[from] UserError),
-
-    #[error("user was not found")]
-    UserNotFound,
-}
-
-/// Handles `UserPictureChangeCommand`.
-pub struct UserPictureChangeCommandHandler<UR>
-where
-    UR: Repository<User>,
-{
-    user_repository: UR,
-}
-
-impl<UR> UserPictureChangeCommandHandler<UR>
-where
-    UR: Repository<User>,
-{
-    pub fn new(user_repository: UR) -> Self {
-        Self { user_repository }
-    }
-}
-
-impl<UR> CommandHandler for UserPictureChangeCommandHandler<UR>
-where
-    UR: Repository<User>,
-{
-    type Command = UserPictureChangeCommand;
-    type Output = UserPictureChangeOutput;
-    type ReplayOutput = UserPictureChangeOutput;
-    type Error = UserPictureChangeCommandHandlerError;
-    type Uow = UR::Uow;
-
-    fn authorization_plan(
-        &self,
-        command: &Self::Command,
-    ) -> Result<AuthorizationPlan, Self::Error> {
-        Ok(AuthorizationPlan::OnlyPrincipals(vec![
-            PrincipalRequirement::AuthenticatedWithRelationship {
-                requirement: RelationshipRequirement::check::<User>(
-                    command.user_id,
-                    UserPictureChangerRelation::REF,
-                ),
-                projector_dependencies: ProjectorDependencies::Some(&[
-                    UserOwnerRelationshipProjectorSpec::DESCRIPTOR,
-                ]),
-            },
-        ]))
-    }
-
-    async fn handle(
-        &self,
-        uow: &mut Self::Uow,
-        request_context: &RequestContext,
-        command: &Self::Command,
-    ) -> Result<CommandHandled<Self::Output, Self::ReplayOutput>, Self::Error> {
-        let Some(mut user) = self.user_repository.find(uow, command.user_id).await? else {
-            return Err(UserPictureChangeCommandHandlerError::UserNotFound);
-        };
-
-        user.change_picture(command.picture.clone())?;
-
-        self.user_repository
-            .save(uow, request_context, &mut user)
-            .await?;
-
-        Ok(CommandHandled::same(UserPictureChangeOutput))
-    }
-}
+pub use user_picture_change_command::UserPictureChangeCommand;
+pub use user_picture_change_command_handler::UserPictureChangeCommandHandler;
+pub use user_picture_change_command_handler_error::UserPictureChangeCommandHandlerError;
+pub use user_picture_change_output::UserPictureChangeOutput;
