+++ examples/banking/server/crates/iam/application/src/command/user/user_picture_change/user_picture_change_command.rs
@@ -1,0 +1,11 @@
+use appletheia::command;
+use banking_iam_domain::{UserId, UserPictureRef};
+use serde::{Deserialize, Serialize};
+
+/// Changes a user's picture.
+#[command(name = "user_picture_change")]
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+pub struct UserPictureChangeCommand {
+    pub user_id: UserId,
+    pub picture: Option<UserPictureRef>,
+}
