+++ examples/banking/server/crates/iam/application/src/command/user/user_picture_change/user_picture_change_output.rs
@@ -1,0 +1,5 @@
+use serde::{Deserialize, Serialize};
+
+/// Returned after a user picture change request is applied.
+#[derive(Clone, Debug, Default, PartialEq, Eq, Serialize, Deserialize)]
+pub struct UserPictureChangeOutput;
